Split hooks from module
