+++ apps/meteor/client/NavBarV2/NavBarPagesToolbar/hooks/useMarketPlaceMenu.tsx
@@ -3,7 +3,7 @@ import type { GenericMenuItemProps } from '@rocket.chat/ui-client';
 import { useTranslation, usePermission, useRouter } from '@rocket.chat/ui-contexts';
 import React from 'react';
 
-import { useUserDropdownAppsActionButtons } from '../../../hooks/useAppActionButtons';
+import { useUserDropdownAppsActionButtons } from '../../../hooks/useUserDropdownAppsActionButtons';
 import { useAppRequestStats } from '../../../views/marketplace/hooks/useAppRequestStats';
 
 export const useMarketPlaceMenu = () => {
